chore(api-cat): Separate QueryUtil from ApiSearchController
