+++ applications/api-cat/src/main/java/no/acat/controller/Common.java
@@ -1,0 +1,5 @@
+package no.acat.controller;
+
+public class Common {
+    public static final String MISSING = "MISSING";
+}
